requirements.txt

langgraph
# python-dotenv
requests
langchain-openai
# streamlit
# langchain
langchain-community
openai
# httpx
tiktoken
streamlit>=1.20.0
langchain>=0.0.187
httpx>=0.24.0
python-dotenv>=1.0.0
langchain_experimental>=0.0.30